refactor: rename "ErrorList" directory.
There is a "ErrorList.cs" file within the "ErrorList" directory, so the directory was renamed and a namespace added to match the folder structure.
