+++ src/LibraryManager.Vsix/Errors/ErrorList.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.LibraryNaming;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Errors
 {
     internal class ErrorList
     {
