+++ src/LibraryManager.Vsix/Errors/TableDataSource.cs
@@ -9,7 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Errors
 {
     internal class TableDataSource : ITableDataSource
     {
