+++ src/LibraryManager.Vsix/Microsoft.Web.LibraryManager.Vsix.csproj
@@ -97,11 +97,11 @@
     <Compile Include="Contracts\HostInteraction.cs" />
     <Compile Include="Contracts\Logger.cs" />
     <Compile Include="Contracts\PerProjectLogger.cs" />
-    <Compile Include="ErrorList\DisplayError.cs" />
-    <Compile Include="ErrorList\ErrorList.cs" />
-    <Compile Include="ErrorList\SinkManager.cs" />
-    <Compile Include="ErrorList\TableDataSource.cs" />
-    <Compile Include="ErrorList\TableEntriesSnapshot.cs" />
+    <Compile Include="Errors\DisplayError.cs" />
+    <Compile Include="Errors\ErrorList.cs" />
+    <Compile Include="Errors\SinkManager.cs" />
+    <Compile Include="Errors\TableDataSource.cs" />
+    <Compile Include="Errors\TableEntriesSnapshot.cs" />
     <Compile Include="Json\Completion\CompletionElementProvider.cs" />
     <Compile Include="Json\Completion\PathCompletionProvider.cs" />
     <Compile Include="Json\Completion\ProviderCompletionProvider.cs" />
@@ -445,7 +445,6 @@
   </Target>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VsSDKInstall)\Microsoft.VsSDK.targets" Condition="Exists('$(VsSDKInstall)\Microsoft.VsSDK.targets')" />
-
   <!--
     Some of the VSSDK packages may pull in Microsoft.VisualStudio.SDK.EmbedInteropTypes which essentially
     does the below but for all of the listed references. We can't just use what they have since they include
